+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/rewrite/AccessPathExpressionCollector.java
@@ -83,15 +83,17 @@ import java.util.Stack;
 public class AccessPathExpressionCollector extends DefaultExpressionVisitor<Void, CollectorContext> {
     private StatementContext statementContext;
     private boolean bottomPredicate;
+    private boolean skipMetaPath;
     private Multimap<Integer, CollectAccessPathResult> slotToAccessPaths;
     private Stack<Map<String, Expression>> nameToLambdaArguments = new Stack<>();
 
     public AccessPathExpressionCollector(
             StatementContext statementContext, Multimap<Integer, CollectAccessPathResult> slotToAccessPaths,
-            boolean bottomPredicate) {
+            boolean bottomPredicate, boolean skipMetaPath) {
         this.statementContext = statementContext;
         this.slotToAccessPaths = slotToAccessPaths;
         this.bottomPredicate = bottomPredicate;
+        this.skipMetaPath = skipMetaPath;
     }
 
     public void collect(Expression expression) {
@@ -203,6 +205,10 @@ public class AccessPathExpressionCollector extends DefaultExpressionVisitor<Void
         // single check covers nested CHAR cases too.
         if (arg.getDataType().isStringLikeType() && !arg.getDataType().isCharType()
                 && context.accessPathBuilder.isEmpty()) {
+            if (skipMetaPath) {
+                return arg.accept(this,
+                        new CollectorContext(context.statementContext, false));
+            }
             CollectorContext offsetContext =
                     new CollectorContext(context.statementContext, context.bottomFilter);
             offsetContext.accessPathBuilder.addSuffix(AccessPathInfo.ACCESS_OFFSET);
@@ -217,6 +223,10 @@ public class AccessPathExpressionCollector extends DefaultExpressionVisitor<Void
         Expression arg = mapSize.child();
         DataType argType = arg.getDataType();
         if (argType.isMapType() && context.accessPathBuilder.isEmpty()) {
+            if (skipMetaPath) {
+                return arg.accept(this,
+                        new CollectorContext(context.statementContext, false));
+            }
             CollectorContext offsetContext =
                     new CollectorContext(context.statementContext, context.bottomFilter);
             offsetContext.accessPathBuilder.addSuffix(AccessPathInfo.ACCESS_OFFSET);
@@ -232,6 +242,10 @@ public class AccessPathExpressionCollector extends DefaultExpressionVisitor<Void
         // Arrays and maps share the same offset-array + data storage layout as strings on the BE.
         DataType argType = arg.getDataType();
         if ((argType.isArrayType() || argType.isMapType()) && context.accessPathBuilder.isEmpty()) {
+            if (skipMetaPath) {
+                return arg.accept(this,
+                        new CollectorContext(context.statementContext, false));
+            }
             CollectorContext offsetContext =
                     new CollectorContext(context.statementContext, context.bottomFilter);
             offsetContext.accessPathBuilder.addSuffix(AccessPathInfo.ACCESS_OFFSET);
@@ -563,6 +577,10 @@ public class AccessPathExpressionCollector extends DefaultExpressionVisitor<Void
         // and nested access (element_at(s, 'city') IS NULL → [s, city, NULL]).
         // For unrecognized expressions, the default visitor resets context, safely discarding NULL.
         if (arg.nullable() && context.accessPathBuilder.isEmpty()) {
+            if (skipMetaPath) {
+                return arg.accept(this,
+                        new CollectorContext(context.statementContext, false));
+            }
             CollectorContext nullContext =
                     new CollectorContext(context.statementContext, context.bottomFilter);
             nullContext.accessPathBuilder.addSuffix(AccessPathInfo.ACCESS_NULL);
