+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/rewrite/AccessPathPlanCollector.java
@@ -64,6 +64,11 @@ import java.util.TreeSet;
 public class AccessPathPlanCollector extends DefaultPlanVisitor<Void, StatementContext> {
     private Multimap<Integer, CollectAccessPathResult> allSlotToAccessPaths = LinkedHashMultimap.create();
     private Map<Slot, List<CollectAccessPathResult>> scanSlotToAccessPaths = new LinkedHashMap<>();
+    private boolean skipMetaPath;
+
+    public void setSkipMetaPath(boolean skipMetaPath) {
+        this.skipMetaPath = skipMetaPath;
+    }
 
     public Map<Slot, List<CollectAccessPathResult>> collect(Plan root, StatementContext context) {
         root.accept(this, context);
@@ -83,7 +88,7 @@ public class AccessPathPlanCollector extends DefaultPlanVisitor<Void, StatementC
         List<Slot> output = generate.getGeneratorOutput();
 
         AccessPathExpressionCollector exprCollector
-                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, false);
+                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, false, skipMetaPath);
         for (int i = 0; i < output.size(); i++) {
             Slot generatorOutput = output.get(i);
             Function function = generators.get(i);
@@ -229,7 +234,7 @@ public class AccessPathPlanCollector extends DefaultPlanVisitor<Void, StatementC
     @Override
     public Void visitLogicalProject(LogicalProject<? extends Plan> project, StatementContext context) {
         AccessPathExpressionCollector exprCollector
-                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, false);
+                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, false, skipMetaPath);
         for (NamedExpression output : project.getProjects()) {
             // e.g. select element_at(s, 'city') from (select s from tbl)a;
             // we will not treat the inner `s` access all path
@@ -385,7 +390,7 @@ public class AccessPathPlanCollector extends DefaultPlanVisitor<Void, StatementC
 
     private void collectByExpressions(Plan plan, StatementContext context, boolean bottomPredicate) {
         AccessPathExpressionCollector exprCollector
-                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, bottomPredicate);
+                = new AccessPathExpressionCollector(context, allSlotToAccessPaths, bottomPredicate, skipMetaPath);
         for (Expression expression : plan.getExpressions()) {
             exprCollector.collect(expression);
         }
