+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/rewrite/NestedColumnPruning.java
@@ -52,7 +52,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Comparator;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -91,9 +90,13 @@ public class NestedColumnPruning implements CustomRewriter {
                 return plan;
             }
             AccessPathPlanCollector collector = new AccessPathPlanCollector();
-            Map<Slot, List<CollectAccessPathResult>> slotToAccessPaths = collector.collect(plan, statementContext);
-            Map<Integer, AccessPathInfo> slotToResult = pruneDataType(slotToAccessPaths,
+            // MV rewrite fragments must not be perturbed by metadata-only paths
+            // (NULL/OFFSET) that would otherwise prune the slot to a narrower type.
+            // Skip collecting them at the source instead of working around them downstream.
+            collector.setSkipMetaPath(
                     jobContext.getCascadesContext().isMaterializedViewRewritePlanFragment());
+            Map<Slot, List<CollectAccessPathResult>> slotToAccessPaths = collector.collect(plan, statementContext);
+            Map<Integer, AccessPathInfo> slotToResult = pruneDataType(slotToAccessPaths);
 
             if (!slotToResult.isEmpty()) {
                 Map<Integer, AccessPathInfo> slotIdToPruneType = Maps.newLinkedHashMap();
@@ -203,8 +206,7 @@ public class NestedColumnPruning implements CustomRewriter {
     }
 
     private static Map<Integer, AccessPathInfo> pruneDataType(
-            Map<Slot, List<CollectAccessPathResult>> slotToAccessPaths,
-            boolean ignoreMetaPath) {
+            Map<Slot, List<CollectAccessPathResult>> slotToAccessPaths) {
         Map<Integer, AccessPathInfo> result = new LinkedHashMap<>();
         Map<Slot, DataTypeAccessTree> slotIdToAllAccessTree = new LinkedHashMap<>();
         Map<Slot, DataTypeAccessTree> slotIdToPredicateAccessTree = new LinkedHashMap<>();
@@ -236,18 +238,6 @@ public class NestedColumnPruning implements CustomRewriter {
                 }
                 continue;
             }
-            if (ignoreMetaPath && containsMetaPath(collectAccessPathResults)) {
-                // An MV rewrite child context optimizes a temporary plan fragment rather
-                // than the final plan. A nested metadata-only path such as
-                // [s, city, NULL] or [s, city, OFFSET] would otherwise prune the scan slot
-                // to only that nested field, while the final MV rewritten plan may still
-                // reuse the same slot as a full complex value or need another child. Drop
-                // access-info for the whole slot instead of just removing that path:
-                // predicate expressions inside this fragment still reference the original
-                // slot shape, so partial pruning after deleting the predicate-only path
-                // could make the fragment itself inconsistent.
-                continue;
-            }
             for (CollectAccessPathResult collectAccessPathResult : collectAccessPathResults) {
                 List<String> path = collectAccessPathResult.getPath();
                 ColumnAccessPathType pathType = collectAccessPathResult.getType();
@@ -273,26 +263,9 @@ public class NestedColumnPruning implements CustomRewriter {
         for (Entry<Slot, DataTypeAccessTree> kv : slotIdToAllAccessTree.entrySet()) {
             Slot slot = kv.getKey();
             DataTypeAccessTree accessTree = kv.getValue();
+            MetaPathNormalizer.normalizeAndStrip(slot, accessTree, allAccessPaths, allAccessPaths);
+
             DataType prunedDataType = accessTree.pruneDataType().orElse(slot.getDataType());
-            normalizeMapValueMetaOnlyAccessPaths(slot, accessTree, allAccessPaths);
-
-            if (accessTree.hasOffsetOnlyAccess() || accessTree.hasNullOnlyAccess()) {
-                if (ignoreMetaPath) {
-                    continue;
-                }
-                stripCoveredMetaPaths(slot, allAccessPaths, allAccessPaths);
-                List<ColumnAccessPath> allPaths = buildColumnAccessPaths(slot, allAccessPaths);
-                result.put(slot.getExprId().asInt(),
-                        new AccessPathInfo(slot.getDataType(), allPaths, new ArrayList<>()));
-                continue;
-            }
-
-            if (slot.getDataType().isStringLikeType()) {
-                // direct access (accessAll=true) or other: skip — no type change, no access paths needed.
-                continue;
-            }
-
-            stripCoveredMetaPaths(slot, allAccessPaths, allAccessPaths);
             List<ColumnAccessPath> allPaths = buildColumnAccessPaths(slot, allAccessPaths);
             if (shouldSkipAccessInfo(slot, prunedDataType, allPaths, predicateAccessPaths)) {
                 continue;
@@ -311,8 +284,7 @@ public class NestedColumnPruning implements CustomRewriter {
         // third: build predicate access path
         for (Entry<Slot, DataTypeAccessTree> kv : slotIdToPredicateAccessTree.entrySet()) {
             Slot slot = kv.getKey();
-            normalizeMapValueMetaOnlyAccessPaths(slot, kv.getValue(), predicateAccessPaths);
-            stripCoveredMetaPaths(slot, predicateAccessPaths, allAccessPaths);
+            MetaPathNormalizer.normalizeAndStrip(slot, kv.getValue(), predicateAccessPaths, allAccessPaths);
             List<ColumnAccessPath> predicatePaths =
                     buildColumnAccessPaths(slot, predicateAccessPaths);
             AccessPathInfo accessPathInfo = result.get(slot.getExprId().asInt());
@@ -336,662 +308,6 @@ public class NestedColumnPruning implements CustomRewriter {
         }
 
         return result;
-    }
-
-    private static boolean containsMetaPath(
-            List<CollectAccessPathResult> collectAccessPathResults) {
-        for (CollectAccessPathResult collectAccessPathResult : collectAccessPathResults) {
-            if (isMetaPath(collectAccessPathResult.getPath())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static boolean isMetaPath(List<String> path) {
-        if (path.isEmpty()) {
-            return false;
-        }
-        String lastComponent = path.get(path.size() - 1);
-        return AccessPathInfo.ACCESS_NULL.equals(lastComponent)
-                || AccessPathInfo.ACCESS_OFFSET.equals(lastComponent);
-    }
-
-    private static void normalizeMapValueMetaOnlyAccessPaths(
-            Slot slot, DataTypeAccessTree accessTree,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> accessPaths) {
-        int slotId = slot.getExprId().asInt();
-        normalizeMapValueMetaPathHelper(slotId,
-                accessTree.collectMapValueMetaOnlyAccessPaths(AccessPathInfo.ACCESS_OFFSET),
-                accessPaths, AccessPathInfo.ACCESS_OFFSET);
-        normalizeMapValueMetaPathHelper(slotId,
-                accessTree.collectMapValueMetaOnlyAccessPaths(AccessPathInfo.ACCESS_NULL),
-                accessPaths, AccessPathInfo.ACCESS_NULL);
-    }
-
-    /**
-     * Normalize map value meta-only (OFFSET or NULL) star access paths for a single meta type.
-     * <p>For each map prefix where {@code [prefix, *, META]} exists, replaces it with
-     * {@code [prefix, KEYS]} plus {@code [prefix, VALUES, META]}, so that keys are read fully
-     * for element lookup while values only read the metadata.
-     */
-    private static void normalizeMapValueMetaPathHelper(
-            int slotId, List<List<String>> mapPrefixes,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> accessPaths,
-            String metaSuffix) {
-        if (mapPrefixes.isEmpty()) {
-            return;
-        }
-
-        Collection<Pair<ColumnAccessPathType, List<String>>> slotPaths = accessPaths.get(slotId);
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToRemove = new ArrayList<>();
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToAdd = new ArrayList<>();
-        for (List<String> mapPrefix : mapPrefixes) {
-            List<String> starMetaPath = new ArrayList<>(mapPrefix);
-            starMetaPath.add(AccessPathInfo.ACCESS_ALL);
-            starMetaPath.add(metaSuffix);
-
-            for (Pair<ColumnAccessPathType, List<String>> p : slotPaths) {
-                if (!p.second.equals(starMetaPath)) {
-                    continue;
-                }
-                pathsToRemove.add(p);
-
-                List<String> keysPath = new ArrayList<>(mapPrefix);
-                keysPath.add(AccessPathInfo.ACCESS_MAP_KEYS);
-                pathsToAdd.add(Pair.of(p.first, keysPath));
-
-                List<String> valuesMetaPath = new ArrayList<>(mapPrefix);
-                valuesMetaPath.add(AccessPathInfo.ACCESS_MAP_VALUES);
-                valuesMetaPath.add(metaSuffix);
-                pathsToAdd.add(Pair.of(p.first, valuesMetaPath));
-            }
-        }
-        slotPaths.removeAll(pathsToRemove);
-        slotPaths.addAll(pathsToAdd);
-    }
-
-    /**
-     * Strip redundant metadata-only NULL/OFFSET paths, keeping enough real paths for BE
-     * readers to avoid OFFSET_ONLY / NULL_MAP_ONLY modes that skip required child data.
-     *
-     * <p>Stripping is organised in two levels:
-     *
-     * <p><b>Level 1 — Same-depth priority:</b> when two paths share the same prefix and
-     * differ only in the final meta suffix, the higher-priority one eliminates the lower.
-     * <pre>{@code
-     *   Data  >  OFFSET  >  NULL
-     * }</pre>
-     * <ul>
-     *   <li>{@code Data} strips {@code OFFSET}: {@code [a]} strips {@code [a, OFFSET]}.</li>
-     *   <li>{@code Data} strips {@code NULL}:  {@code [a]} strips {@code [a, NULL]}.</li>
-     *   <li>{@code OFFSET} strips {@code NULL}: {@code [a, OFFSET]} strips
-     *       {@code [a, NULL]}.</li>
-     * </ul>
-     *
-     * <p><b>Level 2 — Deeper path covers shallower meta:</b> when a covering path goes
-     * deeper into the type tree, its data reader already materialises the container,
-     * making a shallower meta-only path redundant.
-     * <ul>
-     *   <li>Target suffix {@code OFFSET}, covered by deeper:
-     *     <ul>
-     *       <li>{@code Data}:   {@code [a, *, field]}  strips {@code [a, OFFSET]}.</li>
-     *       <li>{@code OFFSET}: {@code [a, *, OFFSET]} strips {@code [a, OFFSET]}.</li>
-     *       <li>{@code NULL}:   {@code [a, *, NULL]}   strips {@code [a, OFFSET]}.</li>
-     *     </ul>
-     *   </li>
-     *   <li>Target suffix {@code NULL}, covered by deeper:
-     *     <ul>
-     *       <li>{@code Data}:   {@code [a, b, c]}      strips {@code [a, b, NULL]}.</li>
-     *       <li>{@code OFFSET}: {@code [a, *, OFFSET]} strips {@code [a, NULL]}.</li>
-     *       <li>{@code NULL}:   {@code [a, *, NULL]}   strips {@code [a, NULL]}.</li>
-     *     </ul>
-     *   </li>
-     * </ul>
-     *
-     * <p>Pre-condition: caller must invoke {@link #normalizeMapValueMetaOnlyAccessPaths}
-     * first to normalize map element-at meta-only access paths (e.g. {@code [m, *, OFFSET]}
-     * becomes {@code [m, KEYS]} + {@code [m, VALUES, OFFSET]}).
-     *
-     * <p>Map OFFSET stripping may need supplemental key paths (e.g.
-     * {@code [m, *, OFFSET]} + {@code [m, VALUES]} becomes {@code [m, KEYS]} +
-     * {@code [m, VALUES]}), handled via {@link #compareOffsetPrefixCoverage} and
-     * {@link #buildMapKeysOnlyPath}. Array NULL has a similar case in
-     * {@link #stripCoveredArrayNullMetaPaths}.
-     */
-    private static void stripCoveredMetaPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> coveringAccessPaths) {
-        // Level 1: same-depth priority — Data > OFFSET > NULL.
-        stripExactPrefixCoveredMetaPaths(slot, targetAccessPaths, coveringAccessPaths);
-        stripNullBySameDepthOffset(slot, targetAccessPaths);
-
-        // Level 2: deeper path covers shallower meta path.
-        stripShallowerOffsetPaths(slot, targetAccessPaths, coveringAccessPaths);
-        stripShallowerNullPaths(slot, targetAccessPaths);
-
-        // Special: array NULL + map key supplementation.
-        stripCoveredArrayNullMetaPaths(slot, targetAccessPaths, coveringAccessPaths);
-    }
-
-    /**
-     * Level 1 — same-depth OFFSET strips NULL: {@code [a, OFFSET]} strips
-     * {@code [a, NULL]}.
-     *
-     * <p>When a NULL and an OFFSET path share the same prefix the OFFSET path is kept
-     * because reading the offset structure already provides null information for
-     * variable-length columns.
-     */
-    private static void stripNullBySameDepthOffset(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> allAccessPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> slotPaths = allAccessPaths.get(slotId);
-        if (slotPaths.isEmpty()) {
-            return;
-        }
-
-        List<Pair<ColumnAccessPathType, List<String>>> toRemove = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : slotPaths) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !AccessPathInfo.ACCESS_NULL.equals(path.get(path.size() - 1))) {
-                continue;
-            }
-            List<String> prefix = path.subList(0, path.size() - 1);
-            for (Pair<ColumnAccessPathType, List<String>> q : slotPaths) {
-                List<String> other = q.second;
-                if (other == path || other.isEmpty()) {
-                    continue;
-                }
-                // Same depth: [a, OFFSET] strips [a, NULL].
-                if (other.size() == path.size()
-                        && AccessPathInfo.ACCESS_OFFSET.equals(other.get(other.size() - 1))
-                        && other.subList(0, prefix.size()).equals(prefix)) {
-                    toRemove.add(p);
-                    break;
-                }
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> r : toRemove) {
-            allAccessPaths.remove(slotId, r);
-        }
-    }
-
-    /**
-     * Decide whether an OFFSET-suffix path can be removed because another path already covers
-     * the same container.
-     *
-     * <p>For map element_at paths, {@code *} means "read keys fully, then follow the rest of
-     * the path on the value side". So a VALUES path can cover the value-side OFFSET access,
-     * but it does NOT cover the key lookup requirement. In that case we remove the OFFSET path
-     * and add a KEYS-only path instead.
-     */
-    private static OffsetPathRewrite analyzeOffsetPathRewrite(
-            DataType slotType, List<String> path, List<List<String>> coveringPaths) {
-        if (path.isEmpty()
-                || !AccessPathInfo.ACCESS_OFFSET.equals(path.get(path.size() - 1))) {
-            return OffsetPathRewrite.keep();
-        }
-        List<String> prefix = path.subList(0, path.size() - 1);
-        // Filter out the path itself to prevent self-coverage removal.
-        // A deeper OFFSET path (e.g. [aa, *, OFFSET]) covering a shallower one
-        // (e.g. [aa, OFFSET]) is still handled — they are different paths.
-        List<List<String>> filteredCoveringPaths = new ArrayList<>();
-        for (List<String> p : coveringPaths) {
-            if (!p.equals(path)) {
-                filteredCoveringPaths.add(p);
-            }
-        }
-        return analyzePrefixCoverage(slotType, prefix, filteredCoveringPaths);
-    }
-
-    private static OffsetPathRewrite analyzePrefixCoverage(
-            DataType slotType, List<String> prefix, List<List<String>> coveringPaths) {
-        List<List<String>> supplementalPaths = new ArrayList<>();
-        for (List<String> coveringPath : coveringPaths) {
-            OffsetPathRewrite candidate = compareOffsetPrefixCoverage(slotType, prefix, coveringPath);
-            if (!candidate.shouldRemoveOffsetPath()) {
-                continue;
-            }
-            if (candidate.getSupplementalPaths().isEmpty()) {
-                return OffsetPathRewrite.remove();
-            }
-            supplementalPaths.addAll(candidate.getSupplementalPaths());
-        }
-        if (supplementalPaths.isEmpty()) {
-            return OffsetPathRewrite.keep();
-        }
-        return OffsetPathRewrite.rewriteWithSupplementalPaths(supplementalPaths);
-    }
-
-    /**
-     * Level 2 — deeper paths cover shallower OFFSET paths:
-     * <ul>
-     *   <li>Deeper {@code Data}:   delegates to {@link #stripCoveredOffsetByPaths}
-     *       for map key supplementation.</li>
-     *   <li>Deeper {@code OFFSET} / {@code NULL}: delegates to
-     *       {@link #stripCoveredMetaByPrefix}.</li>
-     * </ul>
-     */
-    private static void stripShallowerOffsetPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> coveringAccessPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> targetPaths = targetAccessPaths.get(slotId);
-        if (targetPaths.isEmpty()) {
-            return;
-        }
-
-        // Row 1: deeper data paths cover OFFSET (with map key supplementation).
-        List<List<String>> dataPaths = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : coveringAccessPaths.get(slotId)) {
-            List<String> path = p.second;
-            if (!path.isEmpty() && !isMetaPath(path)) {
-                dataPaths.add(path);
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            List<String> path = p.second;
-            if (!path.isEmpty() && !isMetaPath(path)) {
-                dataPaths.add(path);
-            }
-        }
-        stripCoveredOffsetByPaths(slot, targetAccessPaths, dataPaths);
-
-        // Rows 2+3: deeper OFFSET/NULL paths cover shallower OFFSET.
-        // Merge meta paths from covering + target so both inbound and intra-target
-        // coverage are checked.
-        List<List<String>> deeperMetaPaths = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : coveringAccessPaths.get(slotId)) {
-            if (!p.second.isEmpty() && isMetaPath(p.second)) {
-                deeperMetaPaths.add(p.second);
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            if (!p.second.isEmpty() && isMetaPath(p.second)) {
-                deeperMetaPaths.add(p.second);
-            }
-        }
-        stripCoveredMetaByPrefix(slot.getDataType(), slotId, AccessPathInfo.ACCESS_OFFSET,
-                deeperMetaPaths, targetAccessPaths);
-    }
-
-    /**
-     * Level 2 — deeper paths cover shallower NULL paths:
-     * <ul>
-     *   <li>Deeper {@code Data}: any deeper non-meta path whose prefix strictly
-     *       contains the NULL path's prefix strips it.</li>
-     *   <li>Deeper {@code OFFSET} / {@code NULL}: delegates to
-     *       {@link #stripCoveredMetaByPrefix}.</li>
-     * </ul>
-     */
-    private static void stripShallowerNullPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> allAccessPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> slotPaths = allAccessPaths.get(slotId);
-        if (slotPaths.isEmpty()) {
-            return;
-        }
-
-        // Row 1: deeper data paths cover shallower NULL paths.
-        List<Pair<ColumnAccessPathType, List<String>>> toRemove = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : slotPaths) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !AccessPathInfo.ACCESS_NULL.equals(path.get(path.size() - 1))) {
-                continue;
-            }
-            List<String> prefix = path.subList(0, path.size() - 1);
-            for (Pair<ColumnAccessPathType, List<String>> q : slotPaths) {
-                List<String> other = q.second;
-                if (other == path || other.isEmpty() || isMetaPath(other)) {
-                    continue;
-                }
-                // [a] strips [a, NULL]; [a, b, c] strips [a, b, NULL].
-                if (other.equals(prefix) || hasStrictPrefix(other, prefix)) {
-                    toRemove.add(p);
-                    break;
-                }
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> r : toRemove) {
-            allAccessPaths.remove(slotId, r);
-        }
-
-        // Rows 2+3: deeper OFFSET/NULL paths cover shallower NULL paths.
-        List<List<String>> metaPaths = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : slotPaths) {
-            if (!p.second.isEmpty() && isMetaPath(p.second)) {
-                metaPaths.add(p.second);
-            }
-        }
-        stripCoveredMetaByPrefix(slot.getDataType(), slotId, AccessPathInfo.ACCESS_NULL,
-                metaPaths, allAccessPaths);
-    }
-
-    /**
-     * Level 2 — for each target path ending with {@code targetSuffix}, remove it
-     * when a strictly deeper meta path (ending with OFFSET or NULL) has the target
-     * prefix as a strict prefix.
-     *
-     * <p>Both target and covering paths have their meta suffix stripped before
-     * comparison, so only genuinely deeper paths match. Same-depth cross-type
-     * (e.g. {@code [a, OFFSET]} vs {@code [a, NULL]}) is handled by
-     * {@link #stripNullBySameDepthOffset} instead.
-     */
-    private static void stripCoveredMetaByPrefix(
-            DataType slotType, int slotId, String targetSuffix,
-            List<List<String>> coveringMetaPaths,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths) {
-        Collection<Pair<ColumnAccessPathType, List<String>>> targetPaths =
-                targetAccessPaths.get(slotId);
-        if (targetPaths.isEmpty() || coveringMetaPaths.isEmpty()) {
-            return;
-        }
-
-        List<Pair<ColumnAccessPathType, List<String>>> toRemove = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !targetSuffix.equals(path.get(path.size() - 1))) {
-                continue;
-            }
-            List<String> targetPrefix = path.subList(0, path.size() - 1);
-            for (List<String> other : coveringMetaPaths) {
-                if (other == path || other.isEmpty()) {
-                    continue;
-                }
-                List<String> otherPrefix = other.subList(0, other.size() - 1);
-                // Use type-aware comparison so that * ≡ VALUES
-                // (and * ≡ KEYS) at map positions are recognized.
-                OffsetPathRewrite rewrite = compareOffsetPrefixCoverage(
-                        slotType, targetPrefix, otherPrefix);
-                if (rewrite.shouldRemoveOffsetPath()) {
-                    toRemove.add(p);
-                    break;
-                }
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> r : toRemove) {
-            targetAccessPaths.remove(slotId, r);
-        }
-    }
-
-    private static void stripCoveredOffsetByPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths,
-            List<List<String>> coveringPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> targetPaths = targetAccessPaths.get(slotId);
-        if (targetPaths.isEmpty() || coveringPaths.isEmpty()) {
-            return;
-        }
-
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToRemove = new ArrayList<>();
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToAdd = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : new ArrayList<>(targetPaths)) {
-            OffsetPathRewrite rewrite = analyzeOffsetPathRewrite(
-                    slot.getDataType(), p.second, coveringPaths);
-            if (!rewrite.shouldRemoveOffsetPath()) {
-                continue;
-            }
-            pathsToRemove.add(p);
-            for (List<String> supplementalPath : rewrite.getSupplementalPaths()) {
-                pathsToAdd.add(Pair.of(p.first, supplementalPath));
-            }
-        }
-        targetPaths.removeAll(pathsToRemove);
-        targetPaths.addAll(pathsToAdd);
-    }
-
-    /**
-     * Remove array NULL-only paths from {@code targetAccessPaths} when another path already reads
-     * the same array container or data under it. This mirrors OFFSET coverage because an array
-     * element/data read must not be combined with an array NULL_MAP_ONLY read for the same prefix.
-     *
-     * <p>Examples:
-     * <ul>
-     *   <li>{@code [map.VALUES.NULL, map.VALUES.*.field]} becomes
-     *       {@code [map.VALUES.*.field]}.</li>
-     *   <li>{@code [map.*.NULL, map.VALUES.*.field]} becomes
-     *       {@code [map.KEYS, map.VALUES.*.field]} so map lookup keys are still available.</li>
-     * </ul>
-     */
-    private static void stripCoveredArrayNullMetaPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> coveringAccessPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> targetPaths = targetAccessPaths.get(slotId);
-        if (targetPaths.isEmpty()) {
-            return;
-        }
-
-        List<List<String>> nonNullPaths = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : coveringAccessPaths.get(slotId)) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !AccessPathInfo.ACCESS_NULL.equals(path.get(path.size() - 1))) {
-                nonNullPaths.add(path);
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !AccessPathInfo.ACCESS_NULL.equals(path.get(path.size() - 1))) {
-                nonNullPaths.add(path);
-            }
-        }
-
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToRemove = new ArrayList<>();
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToAdd = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : new ArrayList<>(targetPaths)) {
-            List<String> path = p.second;
-            if (path.isEmpty() || !AccessPathInfo.ACCESS_NULL.equals(path.get(path.size() - 1))) {
-                continue;
-            }
-            List<String> prefix = path.subList(0, path.size() - 1);
-            Optional<DataType> prefixType = dataTypeAtPath(slot.getDataType(), prefix);
-            if (!prefixType.isPresent() || !prefixType.get().isArrayType()) {
-                continue;
-            }
-            OffsetPathRewrite rewrite = analyzePrefixCoverage(slot.getDataType(), prefix, nonNullPaths);
-            if (!rewrite.shouldRemoveOffsetPath()) {
-                continue;
-            }
-            pathsToRemove.add(p);
-            for (List<String> supplementalPath : rewrite.getSupplementalPaths()) {
-                pathsToAdd.add(Pair.of(p.first, supplementalPath));
-            }
-        }
-        targetPaths.removeAll(pathsToRemove);
-        targetPaths.addAll(pathsToAdd);
-    }
-
-    /**
-     * Remove exact metadata-only NULL/OFFSET paths when the same field is read in full.
-     * This rule is type-agnostic: once {@code s} itself is accessed, {@code s.NULL} and
-     * {@code s.OFFSET} are redundant and unsafe to keep with the full data path.
-     *
-     * <p>Examples:
-     * <ul>
-     *   <li>{@code [str_col, str_col.NULL]} becomes {@code [str_col]}.</li>
-     *   <li>{@code [arr, arr.OFFSET]} becomes {@code [arr]}.</li>
-     *   <li>{@code [map.*, map.*.OFFSET]} becomes {@code [map.*]}.</li>
-     * </ul>
-     */
-    private static void stripExactPrefixCoveredMetaPaths(
-            Slot slot, Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> targetAccessPaths,
-            Multimap<Integer, Pair<ColumnAccessPathType, List<String>>> coveringAccessPaths) {
-        int slotId = slot.getExprId().asInt();
-        Collection<Pair<ColumnAccessPathType, List<String>>> targetPaths = targetAccessPaths.get(slotId);
-        if (targetPaths.isEmpty()) {
-            return;
-        }
-
-        List<List<String>> fullAccessPaths = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : coveringAccessPaths.get(slotId)) {
-            if (!isMetaPath(p.second)) {
-                fullAccessPaths.add(p.second);
-            }
-        }
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            if (!isMetaPath(p.second)) {
-                fullAccessPaths.add(p.second);
-            }
-        }
-
-        List<Pair<ColumnAccessPathType, List<String>>> pathsToRemove = new ArrayList<>();
-        for (Pair<ColumnAccessPathType, List<String>> p : targetPaths) {
-            List<String> path = p.second;
-            if (!isMetaPath(path)) {
-                continue;
-            }
-            List<String> prefix = path.subList(0, path.size() - 1);
-            for (List<String> fullAccessPath : fullAccessPaths) {
-                if (pathCoversPrefix(fullAccessPath, prefix)) {
-                    pathsToRemove.add(p);
-                    break;
-                }
-            }
-        }
-        targetPaths.removeAll(pathsToRemove);
-    }
-
-    private static Optional<DataType> dataTypeAtPath(DataType slotType, List<String> path) {
-        if (path.isEmpty()) {
-            return Optional.empty();
-        }
-        DataType currentType = slotType;
-        for (int i = 1; i < path.size(); i++) {
-            String component = path.get(i);
-            if (currentType.isStructType()) {
-                StructField field = ((StructType) currentType).getField(component);
-                if (field == null) {
-                    return Optional.empty();
-                }
-                currentType = field.getDataType();
-            } else if (currentType.isArrayType()) {
-                if (!AccessPathInfo.ACCESS_ALL.equals(component)) {
-                    return Optional.empty();
-                }
-                currentType = ((ArrayType) currentType).getItemType();
-            } else if (currentType.isMapType()) {
-                currentType = descendMapType((MapType) currentType, component);
-            } else {
-                return Optional.empty();
-            }
-        }
-        return Optional.of(currentType);
-    }
-
-    private static OffsetPathRewrite compareOffsetPrefixCoverage(
-            DataType slotType, List<String> prefix, List<String> nonOffset) {
-        if (nonOffset.isEmpty()) {
-            return OffsetPathRewrite.remove();
-        }
-        int minLen = Math.min(prefix.size(), nonOffset.size());
-        List<List<String>> supplementalPaths = new ArrayList<>();
-        DataType currentType = slotType;
-        for (int i = 0; i < minLen; i++) {
-            String prefixComponent = prefix.get(i);
-            String nonOffsetComponent = nonOffset.get(i);
-            if (i == 0) {
-                if (!prefixComponent.equals(nonOffsetComponent)) {
-                    return OffsetPathRewrite.keep();
-                }
-                continue;
-            }
-            if (currentType.isStructType()) {
-                if (!prefixComponent.equals(nonOffsetComponent)) {
-                    return OffsetPathRewrite.keep();
-                }
-                StructField field = ((StructType) currentType).getField(prefixComponent);
-                if (field == null) {
-                    return OffsetPathRewrite.keep();
-                }
-                currentType = field.getDataType();
-                continue;
-            }
-            if (currentType.isArrayType()) {
-                if (!prefixComponent.equals(nonOffsetComponent)
-                        || !AccessPathInfo.ACCESS_ALL.equals(prefixComponent)) {
-                    return OffsetPathRewrite.keep();
-                }
-                currentType = ((ArrayType) currentType).getItemType();
-                continue;
-            }
-            if (currentType.isMapType()) {
-                MapType mapType = (MapType) currentType;
-                if (prefixComponent.equals(nonOffsetComponent)) {
-                    currentType = descendMapType(mapType, prefixComponent);
-                    continue;
-                }
-                if (AccessPathInfo.ACCESS_ALL.equals(prefixComponent)
-                        && AccessPathInfo.ACCESS_MAP_VALUES.equals(nonOffsetComponent)) {
-                    supplementalPaths.add(buildMapKeysOnlyPath(prefix, i));
-                    currentType = mapType.getValueType();
-                    continue;
-                }
-                if (AccessPathInfo.ACCESS_MAP_VALUES.equals(prefixComponent)
-                        && AccessPathInfo.ACCESS_ALL.equals(nonOffsetComponent)) {
-                    currentType = mapType.getValueType();
-                    continue;
-                }
-                if (AccessPathInfo.ACCESS_MAP_KEYS.equals(prefixComponent)
-                        && AccessPathInfo.ACCESS_ALL.equals(nonOffsetComponent)) {
-                    currentType = mapType.getKeyType();
-                    continue;
-                }
-                return OffsetPathRewrite.keep();
-            }
-            if (!prefixComponent.equals(nonOffsetComponent)) {
-                return OffsetPathRewrite.keep();
-            }
-        }
-        if (supplementalPaths.isEmpty()) {
-            return OffsetPathRewrite.remove();
-        }
-        return OffsetPathRewrite.rewriteWithSupplementalPaths(supplementalPaths);
-    }
-
-    private static DataType descendMapType(MapType mapType, String component) {
-        if (AccessPathInfo.ACCESS_MAP_KEYS.equals(component)) {
-            return mapType.getKeyType();
-        }
-        return mapType.getValueType();
-    }
-
-    private static List<String> buildMapKeysOnlyPath(List<String> prefix, int mapTokenIndex) {
-        List<String> keyPath = new ArrayList<>(prefix.subList(0, mapTokenIndex));
-        keyPath.add(AccessPathInfo.ACCESS_MAP_KEYS);
-        return keyPath;
-    }
-
-    private static final class OffsetPathRewrite {
-        private static final OffsetPathRewrite KEEP = new OffsetPathRewrite(false, ImmutableList.of());
-        private static final OffsetPathRewrite REMOVE = new OffsetPathRewrite(true, ImmutableList.of());
-
-        private final boolean removeOffsetPath;
-        private final List<List<String>> supplementalPaths;
-
-        private OffsetPathRewrite(boolean removeOffsetPath, List<List<String>> supplementalPaths) {
-            this.removeOffsetPath = removeOffsetPath;
-            this.supplementalPaths = supplementalPaths;
-        }
-
-        private static OffsetPathRewrite keep() {
-            return KEEP;
-        }
-
-        private static OffsetPathRewrite remove() {
-            return REMOVE;
-        }
-
-        private static OffsetPathRewrite rewriteWithSupplementalPaths(List<List<String>> supplementalPaths) {
-            return new OffsetPathRewrite(true, ImmutableList.copyOf(supplementalPaths));
-        }
-
-        private boolean shouldRemoveOffsetPath() {
-            return removeOffsetPath;
-        }
-
-        private List<List<String>> getSupplementalPaths() {
-            return supplementalPaths;
-        }
     }
 
     /**
@@ -1021,14 +337,6 @@ public class NestedColumnPruning implements CustomRewriter {
             }
         }
         predicatePaths.removeAll(toRemove);
-    }
-
-    private static boolean hasStrictPrefix(List<String> path, List<String> prefix) {
-        return path.size() > prefix.size() && path.subList(0, prefix.size()).equals(prefix);
-    }
-
-    private static boolean pathCoversPrefix(List<String> path, List<String> prefix) {
-        return prefix.size() >= path.size() && prefix.subList(0, path.size()).equals(path);
     }
 
     private static List<ColumnAccessPath> buildColumnAccessPaths(
